reorganising and renaming the stuff in these project.
the core will get the name "GECKO3COM_" followed by the type "simple", "plb" or "opb"
to follow the naming in the GECKO3 wiki and to show the IP core interface.

duplicated fifo corecenerator files are merged together including a wrapper to easily supress synthesizer warnings
from unavailable, unused and unconnected pins.

the project is now organised in a way how the IP core and it's parts are beeing used. this means that the 
low-level gpif access module is instantiated by the higher level modules and not the other way around. 
this will make more sense when more parts of this IP core are finished (planning is finished, they have
to be implemented and tested now).