basic synchronous IN (fpga to host) transfer works.
detail changes:
-the scpi command "fpga:data" checks now, if the fpga is configured before it changes the context to the fpga. if the fpga is
not configured, it returns an "EXECUTE ERROR".
in the same way, the main_loop checks if the fpga looses his configuration. it disables the GPIF, switches the context back
to the fx2 if so. this is mainly to avoid undeterministic behaviour if you reconfigure the fpga via jtag.

-introduced the new signal "EOM" end of message from the "usb tmc protokoll interpreter" to the gpif_com module

-changed the GPIF waveform for the FIFO IN transfer to the new scheme.

-implemented the same waveform into the gpif_com_fsm.vhd. works well together.

-bugfixed the gpif_com_test.vhd. sends the new EOM signal, the response message generator works now as it should.

-added the missing AUTHORS README and COPYING (license) files to the core directory.