system is fully working and tested with big files over 4 Gbyte. there were some smaller bugs and glitches which caused sometimes a deadlock. they are fixed now.
one was because the fx2 reads the done pin sometimes wrong during a long data 
transfer. I don't now if this is because of crosstalk on my prototyping system or fpga internal weakness. 
a simple debouncing (read again after 40 ms) did the trick.

changed firmware version to 0.5rc0, basically all features targeted for 0.5 are done.